+++ program/Cargo.toml
@@ -7,7 +7,6 @@ publish = false
 
 [dependencies]
 solana-program = "=1.9.4"
-solana-client = "=1.9.4"
 borsh = "0.9.3"
 bytemuck = {version = "1.7.2", features = ["derive"]}
 spl-token = {version = "3.2.0", features = ["no-entrypoint"]}
@@ -25,6 +24,7 @@ solana-program-test = "=1.9.4"
 solana-sdk = "=1.9.4"
 solana-validator = "=1.9.4"
 solana-logger = "=1.9.4"
+solana-client = "=1.9.4"
 
 [lib]
 crate-type = ["cdylib", "lib"]
