+++ program/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod client;
 pub mod entrypoint;
 pub mod error;
 pub mod instruction;
